Fix a bug in append_gif and move to gif cog
