+++ main.py
@@ -69,7 +69,7 @@ if __name__ == "__main__":
 
     spoof_human()
 
-    cogs = ["color", "transformation", "effect", "draw"]
+    cogs = ["color", "transformation", "effect", "draw", "gif"]
 
     for cog in cogs:
         asyncio.run(bot.load_extension(f"cogs.{cog}_commands"))
